chore: rename changed files to match refactor
